Reorganized node file structure
