+++ NodeBrain/include/NodeBrain/NodeBrain.h
@@ -31,3 +31,8 @@
 #include "Scene/Scene.h"
 #include "Scene/Entity.h"
 #include "Scene/Component.h"
+#include "Scene/EntityGraph.h"
+#include "Scene/Nodes/Node.h"
+#include "Scene/Nodes/TypeNodes.h"
+#include "Scene/Nodes/ComponentNodes.h"
+#include "Scene/Nodes/MathNodes.h"
